requirements.txt
# Core LLM + agent stack
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.3.0

# LangSmith evaluation
langsmith>=0.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
